# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: BANKDATA SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/create:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create crypto payment invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /payments/{id}/status:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Get payment status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentResponse:
      type: object
      properties:
        id:
          type: number
        orderId:
          type: number
        trackId:
          type: string
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
        payAddress:
          type: string
        qrCode:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [id, orderId, amount, currency, status]
    PaymentStatusResponse:
      type: object
      properties:
        id:
          type: number
        status:
          type: string
        confirmedAt:
          type: string
          format: date-time
      required: [id, status]
    CreatePaymentRequest:
      type: object
      properties:
        orderId:
          type: number
        currency:
          type: string
      required: [orderId]
...