# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BANKDATA SaaS Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface PaymentResponse {
  id: number;
  orderId: number;
  trackId?: string;
  amount: string;
  currency: string;
  status: string;
  payAddress?: string;
  qrCode?: string;
  expiresAt?: Date;
}